requirements.txt
amqp==5.0.5
appdirs==1.4.3
asgiref==3.3.1
billiard==3.6.3.0
CacheControl==0.12.6
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.1.6
django-crispy-forms==1.11.0
django-heroku==0.3.1
django-phonenumber-field==5.0.0
django-recaptcha==2.0.6
dnspython==1.16.0
eventlet==0.30.1
greenlet==1.0.0
gunicorn==20.0.4
html5lib==1.0.1
idna==2.10
importlib-metadata==3.4.0
ipaddr==2.2.0
kombu==5.0.2
lockfile==0.12.2
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
phonenumberslite==8.12.18
progress==1.5
prompt-toolkit==3.0.16
psycopg2==2.8.6
pyparsing==2.4.6
pytoml==0.1.21
pytz==2021.1
requests==2.25.1
retrying==1.3.3
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
zipp==3.4.0
